Removed code redundancy, moved server communication out of index, added authentichation to registration and login
