+++ src/Login/Login.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { loginUser, getUsers } from '../index';
-import { useUser } from '../Profile/UserContext';
+import { useUser, getUsers } from '../Profile/UserContext';
 
 const LoginForm = () => {
   const { setUser } = useUser();
@@ -11,6 +10,30 @@ const LoginForm = () => {
   const [emailError, setEmailError] = useState('');
   const [passwordError, setPasswordError] = useState('');
 
+  interface LoginData {
+    username: string;
+    password: string;
+  }
+
+  async function loginUser(logindata: LoginData) {
+    const myHeaders = new Headers({
+      'Content-Type': 'application/json',
+    });
+  
+    const response = await fetch('http://localhost:5000/user/login', {
+      method: 'POST',
+      headers: myHeaders,
+      body: JSON.stringify(logindata)
+    });
+  
+    if (response.ok) {
+      return await response.json();
+    } else {
+      const errorData = await response.json();
+      throw new Error(errorData.message || 'Login failed');
+    }
+  }
+
   function validateEmail(inputEmail: string) {
     const regex = /^\S+@\S+\.\S+$/;
     if(!regex.test(inputEmail)){
@@ -19,61 +42,50 @@ const LoginForm = () => {
     }
     setEmailError('');
     return true;
-}
+  }
 
-function ValidatePassword(inputPassword: string) {
-  const regex = /^(?=.*[a-z])(?=.*\d)[a-zA-Z\d]{8,}$/;
-  if(!regex.test(inputPassword)){
-    setPasswordError('Invalid password format');
-    return false;
-}
-setPasswordError('');
-return true;
-}
+  function ValidatePassword(inputPassword: string) {
+    const regex = /^(?=.*[a-z])(?=.*\d)[a-zA-Z\d]{8,}$/;
+    if(!regex.test(inputPassword)){
+      setPasswordError('Invalid password format');
+      return false;
+  }
+  setPasswordError('');
+  return true;
+  }
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
-
+  
     const payload = {
       username: username.trim(),
       password: password.trim(),
     };
-    console.log(username)
-    console.log(password)
-
+  
     try {
       const response = await loginUser(payload);
-      
-      // Assuming the token is in the response and called 'accessToken'
-      if (response.accessToken) {
-        // Save the access token in localStorage
-        localStorage.setItem('accessToken', response.accessToken);
         
-        try {
-          const userData = await getUsers();
-          
-          if (userData) {
-            setUser(userData); // Save userData in UserContext
-          
-            // Navigate to the /profile page after successfully fetching and setting user data
-            navigate('/profile');
-          } else {
-            // Handle the case when userData is null or undefined
-            throw new Error('No userData received');
-          }
-        } catch (userDataError) {
-          console.error('Could not fetch user data:', userDataError);
-          navigate('/registration');
-          // Handle error by showing user feedback or redirecting
+      // Save the access token in localStorage if it exists in the response
+      if (response.accessToken) {
+        localStorage.setItem('accessToken', response.accessToken);
+        // Fetch the user data
+        const userData = await getUsers();
+        
+        if (userData) {
+          setUser(userData);
+          navigate('/profile');
+        } else {
+          alert('No user data received');
         }
-
       } else {
-        // Handle the case where there is no token in the response
-        console.error('Login succeeded but no access token was returned');
+        alert('Login succeeded but no access token was returned');
       }
     } catch (error) {
-      // Handle any errors that occurred during login
-      console.error(error);
+      if (error instanceof Error) {
+        alert(`Failed to login: ${error.message}. Wrong username or password.`);
+      } else {
+        alert('Failed to login due to an unexpected error.');
+      }
     }
   };
