+++ src/Profile/UserContext.tsx
@@ -1,7 +1,6 @@
 import React, { createContext, useContext, useState, useEffect, ReactNode } from 'react';
-import { getUsers } from '../index';
 
-export interface User {
+interface User {
     userId: string;
     email: string;
     firstName: string;
@@ -28,6 +27,40 @@ type UserContextType = {
   user: User | null;
   setUser: React.Dispatch<React.SetStateAction<User | null>>;
 };
+
+export function getUsers() {
+  // Retrieve the token from localStorage within the function
+  const authtoken = localStorage.getItem('accessToken');
+  
+  if (!authtoken) {
+    throw new Error('Access token is missing');
+  }
+
+  const myHeaders = new Headers({
+    'Content-Type': 'application/json',
+    'Authorization': 'Bearer ' + authtoken
+  });
+
+  return fetch('http://localhost:5000/user', {
+    method: 'GET',
+    headers: myHeaders,
+  })
+  .then(response => {
+    if (response.status === 200) {
+      return response.json(); // if the response is successful, parse the JSON body
+    } else {
+      throw new Error('Something went wrong on the api server!');
+    }
+  })
+  .then(data => {
+    console.debug(data);
+    return data;
+  })
+  .catch(error => {
+    console.error(error);
+    throw error; // Re-throw the error for further handling
+  });
+}
 
 const UserContext = createContext<UserContextType | undefined>(undefined);
